Modularize `getAvailableEarmarks`
