+++ backend/routes/utils/proposal-utils.js
@@ -1,0 +1,23 @@
+function getAvailableEarmarks({ grantsCompleted, projectCategory }) {
+  const availableEarmaks = [];
+  if (grantsCompleted == 0) {
+    if (projectCategory == "outreach")
+      availableEarmaks.push("newprojectoutreach");
+    else availableEarmaks.push("newproject");
+  } else {
+    availableEarmaks.push("general");
+    if (projectCategory == "outreach") availableEarmaks.push("outreach");
+  }
+
+  if (grantsCompleted == 1 || grantsCompleted == 2) {
+    availableEarmaks.push("2nd3rd");
+  }
+
+  availableEarmaks.push("coretech");
+
+  return availableEarmaks;
+}
+
+module.exports = {
+  getAvailableEarmarks,
+};
